requirements.txt
Flask
python-dotenv
requests
random